+++ packages/react/CHANGELOG.md
@@ -1,5 +1,30 @@
 # @reusejs/react
 
+## 6.17.4
+
+### Patch Changes
+
+- 1f264f8: Add props for gallery
+
+## 6.17.3
+
+### Patch Changes
+
+- 7e37ee8: Introduce lazy image loader, file browser, image gallery
+- a269ed2: Make props temp
+
+## 6.17.2
+
+### Patch Changes
+
+- 90b40fc: Add selectedfile callback to file uploader
+
+## 6.17.1
+
+### Patch Changes
+
+- 90bdbea: Add file browser
+
 ## 6.17.0
 
 ### Minor Changes